refactor(write-artifact): rename method
