+++ lib/artifacts/index.js
@@ -2,5 +2,5 @@
 
 module.exports = {
     ArtifactTask: require('./artifact-task'),
-    write: require('./write')
+    writeArtifact: require('./write-artifact')
 };
