+++ lib/artifacts/write-artifact.js
